+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Fixed
-- Compilation error regarding `Cronex.Table` genserver process
+- Adding jobs to the `Cronex.Table` via the `Cronex.Scheduler`
 - `Cronex.Table` ping message handling
 
 ## Version 0.1.0 - 2016/11/02